+++ CHANGELOG.md
@@ -15,6 +15,7 @@ en la distintas versiones de la misma.
 * feat: Seguimiento de actividades en el calendario
 * feat: Los formularios que permiten importar masivamente pueden ahora exportar
 * feat: Cumplimentación del informe del tutor/a laboral
+* feat: Generación del informe del tutor/a laboral
 
 2.2.3 (2020-02-24)
 ------------------